Mocking Toilets (#1323)
* feat: start work with a mocking setup for our cypress tests

* feat: add loading bar and populate from the mock-reports file

* feat: refactors and regen mock json file to include more toilets in London

* feat: try setting up the gh workflow to start the mongo mem

* feat: try cachine areainfo
