+++ .env.test
@@ -1,1 +1,2 @@
 AUTH0_BASE_URL="http://localhost:3000"
+MONGODB_URI=mongodb://localhost:27017/toiletmap
