+++ .github/workflows/cypress-e2e.yml
@@ -12,10 +12,18 @@ jobs:
         with:
           node-version-file: '.nvmrc'
           cache: 'yarn'
+      - name: CacheAreas
+        id: cache-areas
+        uses: actions/cache@v3
+        with:
+          path: ./src/api/db/manage/areaToDatabase/.areaInfo.geojson
+          key: ${{ runner.os }}-areaInfo
+      - run: yarn
+      - run: yarn startMemoryMongo & yarn wait-on:memoryMongo
       - name: Cypress run
         uses: cypress-io/github-action@v2
         env:
-          MONGODB_URI: ${{ secrets.MONGODB_URI }}
+          MONGODB_URI: mongodb://localhost:27017/toiletmap
           DATABASE_NAME: toiletmap
           AUTH0_SECRET: ${{ secrets.AUTH0_SECRET }}
           AUTH0_ISSUER_BASE_URL: https://gbptm.eu.auth0.com
