+++ cypress/integration/index.spec.ts
@@ -15,15 +15,14 @@ describe('Home page tests', () => {
     cy.findByText(
       'Hammersmith, Greater London, England, W6 9YA, United Kingdom'
     ).click();
-    cy.get('[data-toiletid="cb92afde0b7ac9483e5d9639"]').should('exist');
+    cy.get('[data-toiletid=891ecdfaf8d8e4ffc087f7ce]').should('exist');
   });
 
   it('should load different toilets when the map is dragged', () => {
     // Drag the map and check that we have fewer markers than when we started
     cy.visit('/').wait(500);
-    cy.get('.toilet-marker').should('have.length.above', 50);
-    cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').should('exist');
-    cy.get('[data-toiletid=1dd0f403d1694d6f7dd33f9a]').should('not.exist');
+    cy.get('[data-toiletid=ddad1ed1b91d99ed2bf3bcdf]').should('exist');
+    cy.get('[data-toiletid=cc4e5e9b83de8dd9ba87b3eb]').should('not.exist');
     cy.get('#gbptm-map')
       .trigger('mousedown', { which: 1 })
       .trigger('mousemove', { which: 1, x: 900, y: 0 })
@@ -34,29 +33,28 @@ describe('Home page tests', () => {
       .trigger('mouseup')
       .wait(100)
       .trigger('mousedown', { which: 1 })
-      .trigger('mousemove', { which: 1, x: 900, y: 0 })
+      .trigger('mousemove', { which: 1, x: 500, y: 0 })
       .trigger('mouseup')
-      .wait(500);
-    cy.get('.toilet-marker').should('have.length.below', 40);
+      .trigger('wheel', { deltaY: 10 });
     cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').should('not.exist');
-    cy.get('[data-toiletid=1dd0f403d1694d6f7dd33f9a]').should('exist');
+    cy.get('[data-toiletid=cc4e5e9b83de8dd9ba87b3eb]').should('exist');
   });
 
   it('should open the toilet details panel when a marker is clicked', () => {
     cy.visit('/').wait(500);
-    cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').click();
+    cy.get('[data-toiletid=ddad1ed1b91d99ed2bf3bcdf]').click({ force: true });
 
-    cy.url().should('include', '/loos/51f6b4d8b792e3531efe5152');
+    cy.url().should('include', '/loos/ddad1ed1b91d99ed2bf3bcdf');
 
     // Check that the loo we picked is now highlighted.
     cy.get('#highlighted-loo').invoke(
       'attr',
       'data-toiletid',
-      '51f6b4d8b792e3531efe5152'
+      'ddad1ed1b91d99ed2bf3bcdf'
     );
 
     // Check standard loo panel stuff is there.
-    cy.contains('Test loo 2.1');
+    cy.contains('negative eve');
     cy.contains('Features');
     cy.contains('Opening Hours');
 
@@ -72,12 +70,12 @@ describe('Home page tests', () => {
 
   it('should collapse the toilet panel when the close button is clicked and reopen when details is clicked', () => {
     cy.visit('/').wait(500);
-    cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').click();
+    cy.get('[data-toiletid=ab2ebfbdadb963aed4cb3b65]').click({ force: true });
     cy.get('[aria-label="Close toilet details"]').click();
 
-    cy.url().should('include', '/loos/51f6b4d8b792e3531efe5152');
+    cy.url().should('include', '/loos/ab2ebfbdadb963aed4cb3b65');
 
-    cy.contains('Test loo 2.1');
+    cy.contains('guilty illiteracy');
     cy.contains('Features').should('not.exist');
     cy.contains('Opening Hours').should('not.exist');
 
@@ -94,39 +92,39 @@ describe('Home page tests', () => {
 
   it('should close the toilet panel fully when the close button is clicked twice', () => {
     cy.visit('/');
-    cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').click();
-    cy.url().should('include', '/loos/51f6b4d8b792e3531efe5152');
+    cy.get('[data-toiletid=ab2ebfbdadb963aed4cb3b65]').click({ force: true });
+    cy.url().should('include', '/loos/ab2ebfbdadb963aed4cb3b65');
     cy.get('[aria-label="Close toilet details"]').scrollIntoView().click();
     cy.contains('Close').scrollIntoView().click();
-    cy.url().should('not.include', '/loos/51f6b4d8b792e3531efe5152');
+    cy.url().should('not.include', '/loos/ab2ebfbdadb963aed4cb3b65');
   });
 
   it('should close the toilet panel when the esc key is pressed', () => {
     cy.visit('/').wait(500);
-    cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').click();
-    cy.url().should('include', '/loos/51f6b4d8b792e3531efe5152');
-    cy.contains('Test loo 2.1');
+    cy.get('[data-toiletid=ab2ebfbdadb963aed4cb3b65]').click({ force: true });
+    cy.url().should('include', '/loos/ab2ebfbdadb963aed4cb3b65');
+    cy.contains('guilty illiteracy');
     cy.contains('Features');
     cy.contains('Opening Hours');
     cy.get('body').trigger('keydown', { key: 'Escape' });
-    cy.contains('Test loo 2.1');
+    cy.contains('guilty illiteracy');
     cy.contains('Features').should('not.exist');
     cy.contains('Opening Hours').should('not.exist');
     cy.get('body').trigger('keydown', { key: 'Escape' });
-    cy.url().should('not.include', '/loos/51f6b4d8b792e3531efe5152');
+    cy.url().should('not.include', '/loos/ab2ebfbdadb963aed4cb3b65');
   });
 
   it('should go to the edit page if the toilet edit button is clicked', () => {
     cy.visit('/').wait(500);
-    cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').click();
+    cy.get('[data-toiletid=ab2ebfbdadb963aed4cb3b65]').click({ force: true });
     cy.findByText('Edit').scrollIntoView().click();
-    cy.url().should('include', 'loos/51f6b4d8b792e3531efe5152/edit');
+    cy.url().should('include', 'loos/ab2ebfbdadb963aed4cb3b65/edit');
     cy.contains('Want to Contribute Toilet Data?');
   });
 
   it('should open directions if the directions button is clicked', () => {
     cy.visit('/').wait(500);
-    cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').click();
+    cy.get('[data-toiletid=ab2ebfbdadb963aed4cb3b65]').click({ force: true });
 
     cy.findByText('Directions')
       .scrollIntoView()
@@ -136,7 +134,7 @@ describe('Home page tests', () => {
 
   it('should allow users to confirm that the toilet information is correct', () => {
     cy.visit('/').wait(500);
-    cy.get('[data-toiletid=51f6b4d8b792e3531efe5152]').click();
+    cy.get('[data-toiletid=ab2ebfbdadb963aed4cb3b65]').click({ force: true });
     cy.intercept('POST', '/api', (req) => {
       expect(req.body.operationName).to.equal(
         'submitVerificationReportMutation'
