+++ src/components/MapState.tsx
@@ -2,7 +2,6 @@ import { Map } from 'leaflet';
 import React, { Dispatch, useEffect } from 'react';
 import { Loo } from '../api-client/graphql';
 import config, { Filters, FILTERS_KEY } from '../config';
-// import { CompressedLooObject } from '../lib/loo';
 
 import { UseLocateMapControl } from './LooMap/useLocateMapControl';
